requirements.txt
# Core FastAPI and web framework dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database dependencies
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
aiosqlite==0.19.0

# Redis and caching
redis==5.0.1
aioredis==2.0.1

# HTTP client libraries
httpx==0.25.2
aiohttp==3.9.1
requests==2.31.0

# Authentication and security
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
cryptography==41.0.8

# AutoGen and AI dependencies
autogen-core==0.2.0
openai==1.3.7
azure-openai==1.3.0

# Monitoring and metrics
prometheus-client==0.19.0
psutil==5.9.6

# Logging and utilities
structlog==23.2.0
coloredlogs==15.0.1
python-json-logger==2.0.7

# Configuration and environment
python-dotenv==1.0.0
click==8.1.7
typer==0.9.0

# Testing dependencies (development)
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2  # For testing HTTP clients
factory-boy==3.3.0
faker==20.1.0

# Code quality and formatting (development)
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0

# Documentation (development)
mkdocs==1.5.3
mkdocs-material==9.4.8
mkdocstrings[python]==0.24.0

# Additional utilities
python-dateutil==2.8.2
pytz==2023.3
tenacity==8.2.3
jinja2==3.1.2
markupsafe==2.1.3

# Development and debugging
ipython==8.17.2
ipdb==0.13.13
rich==13.7.0

# File handling and validation
pillow==10.1.0
validators==0.22.0

# Async utilities
anyio==4.1.0
trio==0.23.1

# Process management
supervisor==4.2.5

# Environment-specific dependencies
# Production
gevent==23.9.1
eventlet==0.33.3

# Optional: Docker health checks
waitress==2.1.2